refactor(do): extract role-split relay tier from ShardDO
The relay transport (~480 lines of state + methods) lived directly in the
8.3k-line ShardDO with owner-only and relay-only state separated only by
comments. It now lives in relay-hub.ts, split by role the way a DO actually
is — an OwnerRelay (relay set, uniform/proxy registries, the RLS-uniform
gate, cohort multicast + per-socket proxy, seed builder) and a RelayMember
(per-socket cohort memos, announce latch, seed-via-owner, poke delivery),
over a shared RelayLink base (addressing + whisper local-delivery + the
control-channel dispatch). The role is chosen ONCE from the DO name at
construction, so owner-only state can never sit next to relay-only state.

ShardDO reaches the tier through a narrow RelayHost seam (shape resolution,
op-log diff/seed, socket set, shared poke-id + fan-out counters) and holds a
single `relay` collaborator it delegates flush/seed/whisper/control to —
undefined (inert) on an unnamed single-DO. ResolvedShape/SubscriptionIdentity
moved to the neutral types.ts and awaitWsDrain to subscription-delivery so
relay-hub has zero dependency on shard-do (no cycle). shard-do.ts drops from
8325 to 7393 lines. Pure structural change — 35 relay tests + the full 981
suite pass unchanged, typecheck + lint clean, runtime unaffected.
